medical_LLM-main/requirements.txt
gradio
python-dotenv
groq
speechrecognition
gtts
torch
torchvision
Pillow
# 📁 requirements.txt -----
flask>=2.0.3
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1